fix broken CA certificates by including the required CA extensions
Previously several applications/libraries would complain that the CA
is invalid due to the missing CA:TRUE attribute.

It is *assumed* that the previous behaviour was intentional, because
creating a CA cert with certutil expects user input, which is not
possible when running this command with Puppet.

This commit uses openssl to create CA certificates, which involves
more steps, but reliably produces a fully functional CA.